Homepage componet imlented with new Header
